// src/lib/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserRole = {
  admin: 'admin',
  head: 'head',
  closer: 'closer',
  sdr: 'sdr',
  operational: 'operational'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const WorkLocation = {
  office: 'office',
  home: 'home',
  day_off: 'day_off'
} as const

export type WorkLocation = (typeof WorkLocation)[keyof typeof WorkLocation]


export const PaymentMethod = {
  pix: 'pix',
  card: 'card',
  boleto: 'boleto'
} as const

export type PaymentMethod = (typeof PaymentMethod)[keyof typeof PaymentMethod]


export const SaleOrigin = {
  organic: 'organic',
  referral: 'referral',
  outbound: 'outbound',
  advance: 'advance'
} as const

export type SaleOrigin = (typeof SaleOrigin)[keyof typeof SaleOrigin]


export const RevenueTier = {
  small: 'small',
  medium: 'medium',
  large: 'large',
  enterprise: 'enterprise'
} as const

export type RevenueTier = (typeof RevenueTier)[keyof typeof RevenueTier]